+++ app/Portal.Consultoras.ServiceContracts/IPedidoService.cs
@@ -1250,6 +1250,6 @@ namespace Portal.Consultoras.ServiceContracts
         bool UpdCantidadPedidoWebSet(int paisID, int SetId, int Cantidad);
 
         [OperationContract]
-        IList<BEPedidoWebDetalle> GetPedidoWebSetProducto(int paisID, int Campania, int ConsultoraId, int Cantidad);
+        IList<BEPedidoWebDetalle> GetPedidoWebSetProducto(int paisID, int Campania, long ConsultoraId, int Cantidad);
     }
 }
